@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class BazarApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(BazarApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(BazarApplication.class, args);
+    }
 
 }
