@@ -1,4 +1,4 @@
-package com.api.bazar.restcontroller;
+package com.api.bazar.controller.api;
 
 import com.api.bazar.dto.VentaDTO;
 import com.api.bazar.model.Producto;
@@ -7,22 +7,26 @@ import com.api.bazar.model.Venta;
 import com.api.bazar.service.IVentaService;
 import java.util.ArrayList;
 import java.util.stream.Collectors;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class VentaRestController {
+@RequestMapping("/api/venta")
+public class VentaApiController {
 
-    @Autowired
-    private IVentaService ventaService;
+    private final IVentaService ventaService;
 
-    @GetMapping("/venta/verVentas")
+    public VentaApiController(IVentaService ventaService) {
+        this.ventaService = ventaService;
+    }
+
+    @GetMapping
     public List<VentaDTO> verVentas() {
         List<Venta> ventas = ventaService.findAll();
         List<VentaDTO> ventasDTO = new ArrayList<>();
@@ -45,7 +49,7 @@ public class VentaRestController {
         return ventasDTO;
     }
 
-    @GetMapping("/venta/verVenta/{id}")
+    @GetMapping("/{id}")
     public VentaDTO verVenta(@PathVariable Long id) {
         Venta vent = ventaService.findById(id);
         VentaDTO ventDTO = new VentaDTO();
@@ -64,19 +68,19 @@ public class VentaRestController {
         return ventDTO;
     }
 
-    @PostMapping("/venta/crearVenta")
+    @PostMapping
     public String crearVenta(@RequestBody Venta venta) {
         ventaService.saveOne(venta);
         return "Venta guardado correctamente";
     }
 
-    @PutMapping("/venta/editarVenta/{id}")
+    @PutMapping("/{id}")
     public String editarVenta(@PathVariable Long id, @RequestBody Venta venta) {
         ventaService.updateOne(venta, id);
         return "Venta actualizado correctamente";
     }
 
-    @DeleteMapping("/venta/eliminarVenta/{id}")
+    @DeleteMapping("/{id}")
     public String eliminarVenta(@PathVariable Long id) {
         ventaService.deleteOne(id);
         return "Venta eliminado con exito";
