@@ -1,9 +1,8 @@
-package com.api.bazar.controller;
+package com.api.bazar.controller.web;
 
 import com.api.bazar.model.Cliente;
 import com.api.bazar.service.IClienteService;
 import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,8 +14,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @Controller
 public class ClienteController {
 
-    @Autowired
-    private IClienteService clienteService;
+    private final IClienteService clienteService;
+
+    public ClienteController(IClienteService clienteService) {
+        this.clienteService = clienteService;
+    }
 
     @RequestMapping("/cliente")
     public String verPaginaCliente(Model model) {
