@@ -1,4 +1,4 @@
-package com.api.bazar.controller;
+package com.api.bazar.controller.web;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
