@@ -1,9 +1,8 @@
-package com.api.bazar.controller;
+package com.api.bazar.controller.web;
 
 import com.api.bazar.model.Producto;
 import com.api.bazar.service.IProductoService;
 import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,8 +14,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @Controller
 public class ProductoController {
 
-    @Autowired
-    private IProductoService prodServi;
+    private final IProductoService prodServi;
+
+    public ProductoController(IProductoService prodServi) {
+        this.prodServi = prodServi;
+    }
 
     @RequestMapping("/producto")
     public String verPaginaProducto(Model model) {
