@@ -3,14 +3,16 @@ package com.api.bazar.service;
 import com.api.bazar.model.Cliente;
 import com.api.bazar.repository.IClienteRepository;
 import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 public class ClienteService implements IClienteService {
 
-    @Autowired
-    private IClienteRepository cliRepo;
+    private final IClienteRepository cliRepo;
+
+    public ClienteService(IClienteRepository cliRepo) {
+        this.cliRepo = cliRepo;
+    }
 
     @Override
     public List<Cliente> findAll() {
@@ -32,7 +34,7 @@ public class ClienteService implements IClienteService {
     @Override
     public void updateOne(Cliente cli, Long id) {
         Cliente cliente = this.findById(id);
-        
+
         cliente.setNombre(cli.getNombre());
         cliente.setApellido(cli.getApellido());
         cliente.setDni(cli.getDni());
